feat: add example app home screen
