+++ example/src/app/(home)/_layout.tsx
@@ -1,0 +1,69 @@
+import { Stack } from 'expo-router';
+import { useTheme } from 'heroui-native';
+import { useCallback } from 'react';
+import { Image, Platform, StyleSheet } from 'react-native';
+import { ThemeToggle } from '../../components/theme-toggle';
+
+export default function Layout() {
+  const { theme, colors } = useTheme();
+
+  const _renderTitle = () => {
+    return (
+      <Image
+        source={{
+          uri: 'https://raw.githubusercontent.com/heroui-inc/heroui/main/apps/docs/public/isotipo.png',
+        }}
+        style={styles.logo}
+      />
+    );
+  };
+
+  const _renderThemeToggle = useCallback(() => <ThemeToggle />, []);
+
+  return (
+    <Stack
+      screenOptions={{
+        headerShown: false,
+        headerTitleAlign: 'center',
+        headerTransparent: Platform.select({
+          ios: true,
+          android: false,
+        }),
+        headerBlurEffect: theme === 'dark' ? 'dark' : 'light',
+        headerTintColor: colors.foreground,
+        headerStyle: {
+          backgroundColor: Platform.select({
+            ios: undefined,
+            android: colors.background,
+          }),
+        },
+        headerTitleStyle: {
+          fontFamily: 'Inter_600SemiBold',
+        },
+        headerRight: _renderThemeToggle,
+        headerBackButtonDisplayMode: 'minimal',
+        gestureEnabled: true,
+        gestureDirection: 'horizontal',
+        fullScreenGestureEnabled: true,
+      }}
+    >
+      <Stack.Screen
+        name="index"
+        options={{ headerShown: true, headerTitle: _renderTitle }}
+      />
+      <Stack.Screen name="components" />
+      <Stack.Screen
+        name="themes/index"
+        options={{ headerShown: true, headerTitle: 'Themes' }}
+      />
+      <Stack.Screen name="showcases" />
+    </Stack>
+  );
+}
+
+const styles = StyleSheet.create({
+  logo: {
+    width: 24,
+    height: 24,
+  },
+});
