+++ example/src/app/(home)/showcases/index.tsx
@@ -1,0 +1,5 @@
+import { View } from 'react-native';
+
+export default function App() {
+  return <View />;
+}
