+++ src/components/drop-shadow-view/drop-shadow-view.styles.ts
@@ -4,7 +4,7 @@ import { tv } from 'tailwind-variants';
  * Main DropShadow component styles
  */
 const root = tv({
-  base: '',
+  base: 'bg-background',
 });
 
 export default root;
